Implemented automatic job price calculation, default value in table of orders when the price is 0, new endpoint, added clickable link's to tables

+++ webAPI/webAPI.Domain/DTOs/AddJobEquipmentDto.cs
@@ -17,7 +17,7 @@ namespace webAPI.Domain.DTOs
 
         public EquipmentType? EquipmentType { get; set; }
 
-		public int CostPerHour { get; set; }
+		public decimal CostPerHour { get; set; }
     }
 }
